fix: change native query in repo's with Pageable interface

+++ main_svc/src/main/java/ru/practicum/main/controllers/admin/AdminCategoryController.java
@@ -30,7 +30,7 @@ public class AdminCategoryController {
     }
 
     @PatchMapping("/{catId}")
-    public CategoryDto updateCategory(@PathVariable Integer catId, @Valid @RequestBody CategoryDto categoryDto) {
+    public CategoryDto updateCategory(@PathVariable Long catId, @Valid @RequestBody CategoryDto categoryDto) {
         return categoryService.updateCategory(catId, categoryDto);
     }
 }
